requirements.txt
eli5
librosa
matplotlib
numpy
pandas
scikeras
scikit-learn
seaborn
tensorflow
tqdm
xgboost